IRIS: Support binary relocation
This patch allows binary relocation when the FILESDIR macro is a relative path.
A second macro BINDIR is used in combination to indicate the (optional) binary
sub folder where the program is installed (e.g. the bin part in /usr/bin).

When FILESDIR is a relative path, the computed files dir will be the concatenation of
the executable parent path [- <BINDIR>] + <FILESDIR>. If FILESDIR is an absolute path,
it is used unmodified.